Store cell content in a global dictionary
Probably not the most efficient way to store it, but it works for now.
A 'keyup' callback is registered against the cell when a user clicks in
it.  If a user edits the cell, it's highlighted red to indicate it's
been changed.  If the user presses enter after making a change,
the cell value is written to global map/dict; keyed on the cell's id.
The border is also changed back to black as a visual indicator that
the value has been stashed.