Store a reference to cells in an equation.
For example, if cell 'A4' contains the equation 'A1 + B1', we now update
the global map (gReferences) with the values:

    { 'A1' => {'A4'}, 'B1' => {'A4'}}

So, the cells in the equation point to a set of cells that use them.
This is for use to use later, so we can update any equations if a
referenced cell is updated.